Add support for hairpin NAT
Currently only flows coming from the upstream interface are DNATed for port
forwarding. When a VM tries to connect to a forwarded port on our public
IPv4 the flow doesn't match and goes nowhere.

To "fix" this we change the prerouting rules to DNAT any --dst-type LOCAL
flow and also add a SNAT rule to POSTROUTING for flows from kpu.

This does have the downside that connections from other VMs look like
they're coming from 10.0.0.1, but I'll accept that for now to have it work
at least.