Implement PVLAN separation of VMs and fix hairpin nat
We've wanted to properly isolate VMs from spoofing on the network for a
long time now. I've recently found this technique known (by some) as
"Private VLAN".

Turns out linux has implemented support for this in the bridge since 2018
with the "isolated on" flag bridge-port flag. Once we turn this on
communication between VMs will break though. To get it back up we have to
get the VMs to route subnet local addresses through the router/hypervisor.

To do this we use two different techniques v4/v6. In v4 land we need PVLAN
Proxy ARP, a special case of regular Proxy ARP where the ARPs proxied back
out the same interface. Linux thankfully also had support for this for a
while now we just have to enable the proxy_arp_pvlan sysctl for our bridge
interface.

In v6 land we can simply modify the NDP router advertisements to not have
the "on-link" flag set in the prefix's advert. This will direct hosts to
send packets to the link-local address of the router instead. Neat.

There are a couple of other leavers we need to flip to make everything
work. By default linux sends ICMP(v6) redirects when it routes packets it
knows are on-link. This breaks connectivity after the first packet though
as hosts will then try to talk directly via L2 again.

To prevent this in the v4 case we just set the send_recirects sysctl
appropriately. In v6 land this setting is missing and we drop these packets
in the iptables OUTPUT chain instead.

Now finally we can fix another peice of nastiness in our network setup. To
support hairpin nat we previously had a SNAT rule that makes flows from
internal hosts towards port forwarded ports on the public IP look like
their coming from the router/hypervisor so we can NAT the reply packets.

Now that eveything is going via the router anyway this is no longer
necessary and just the DNAT rules work just fine.

While we're at it We also prevent the DNAT rules for firing for the
hypervisor IP as this prevents DNS connectivity to dnsmasq when there's
port forwardings for port 53.