Fix hairpin-hairpin NAT
With the new PVLAN setup we don't need the hairpin SNAT rule anymore, well
for most connections. If a VM ends up accessing a port forward pointing to
itself it'll get a reply packet coming from itself which is promtly
delivers without routing it to the route. This is just like before except
now the problem only exists with connections from one VM which makes me
less repulsed by loosing source information as we can be fairly sure the
hypervisor isn't attacking hosts.